integration/app/requirements.txt
alembic==1.11.1
APScheduler==3.10.1
async-timeout==4.0.2
atlassian-python-api==3.36.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
Deprecated==1.2.14
dnspython==2.3.0
email-validator==1.3.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
gevent==23.7.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multi-key-dict==2.0.3
oauthlib==3.2.2
pbr==5.11.1
pyartifactory==1.12.0
pydantic==1.10.11
PyMySQL==1.1.0
python-gitlab==2.10.1
python-jenkins==1.8.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
SQLAlchemy==2.0.18
tencentcloud-sdk-python==3.0.821
tencentcloud-sdk-python-sms==3.0.821
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
zope.event==5.0
zope.interface==6.0
